Cleaned latex and updated graphs
I cleaned up some of the prose in the .tex file and fixed a few typos,
but no major revisions. I've reworked the code for the first graph,
mainly to change the colours, but I also ended up getting rid of a
nasty bug that gave the wrong names to some of the features - best
cross-check the graph matches our text. More to come, I'm going to bed
now.